+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/environment-check
 
+## 0.17.33
+
+### Patch Changes
+
+-   Updated dependencies [df29368]
+    -   @sap-ux/project-access@1.27.0
+    -   @sap-ux/axios-extension@1.16.5
+
 ## 0.17.32
 
 ### Patch Changes