+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 1.0.21
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.3.18
+-   @sap-ux/odata-service-writer@0.22.4
+-   @sap-ux/ui5-application-writer@1.1.6
+
 ## 1.0.20
 
 ### Patch Changes