+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/ui-components
 
+## 1.17.9
+
+### Patch Changes
+
+-   ea0674c: UIContextualMenu.getUIcontextualMenuCalloutStyles - change parameter from `props?: IContextualMenuProps` to `styles?: IStyleFunctionOrObject<IContextualMenuStyleProps, IContextualMenuStyles>` and make parameter as optional
+
+## 1.17.8
+
+### Patch Changes
+
+-   b124873: UIContextualMenu.getUIContextualMenuItemStyles - make `props` param optional
+
 ## 1.17.7
 
 ### Patch Changes